+++ source/assets/components/ViewMyRecipe.js
@@ -37,7 +37,8 @@ class ViewMyRecipe extends HTMLElement {
             flex-flow: row wrap;
             margin: 10px 0px 90px 0px;
         }
-        
+
+
         /* header */
         article > header {
             align-items: center;
@@ -47,7 +48,6 @@ class ViewMyRecipe extends HTMLElement {
             height: 12vh;
             justify-content: space-between;
             width: 100%;
-            border-style:dotted;
         }
 
         #logo-sec {
@@ -58,7 +58,7 @@ class ViewMyRecipe extends HTMLElement {
 
         #logo-sec > a > img {
             height: 100%;
-            width: 100px;
+            width: 125px;
             object-fit: cover;
         }
 
@@ -82,25 +82,9 @@ class ViewMyRecipe extends HTMLElement {
         #home-link {
             color: white;
             font-size: 2vw;
-            // margin: 0px 0px 0px 20px;
+            margin: 0px 0px 0px 20px;
             text-decoration: none;
         }
-
-        @media (max-width: 480px) {
-            article {
-                box-shadow: 0px 0px 15px #888888;
-                display: flex;
-                flex-flow: row wrap;
-                margin: 10px 0px 90px 0px;
-            }
-
-            #header-title {
-                font-size: 5vw; 
-                position:relative;
-                left: 6vw;
-            }
-
-          }
 
 
         /* main */
@@ -108,7 +92,7 @@ class ViewMyRecipe extends HTMLElement {
             align-content: flex-start;
             display: flex;
             flex: 1 1 70%;
-            flex-flow: column wrap;
+            flex-flow: row wrap;
             justify-content: center;
             padding: 15px;
         }
@@ -117,11 +101,10 @@ class ViewMyRecipe extends HTMLElement {
         #main-header {
             align-items: center;
             display: flex;
-            flex-flow: column nowrap;
+            flex-flow: row nowrap;
             justify-content: center;
             margin: 20px 0px 50px 0px;
             width: 100%;
-            text-align: center;
         }
 
         #main-header > h1 {
@@ -131,20 +114,13 @@ class ViewMyRecipe extends HTMLElement {
             margin: 0px 30px;
         }
 
-        @media (max-width: 480px) {
-            article > main {
-                width: calc(100vw - 480px);
-            }
-        }
-
 
         /* left-main */
         #left-main {
             align-items: center;
             display: flex;
-            // flex: 1 1 40%;
+            flex: 1 1 40%;
             flex-flow: column nowrap;
-            width: 100%;
         }
 
         #left-main > img {
@@ -154,13 +130,13 @@ class ViewMyRecipe extends HTMLElement {
 
         #left-main > h2 {
             visibility: hidden;
-            // margin: 10px 0px 20px 0px;
+            margin: 10px 0px 20px 0px;
         }
         
         #left-main > p {
             visibility: hidden;
             font-size: 20px;
-            // margin-bottom: 10px;
+            margin-bottom: 10px;
         }
 
         #left-main > button {
@@ -203,12 +179,6 @@ class ViewMyRecipe extends HTMLElement {
             margin: 20px 5px 20px 30px;
             text-indent: 0px;
         }
-
-        #steps-list > ol > li:hover {
-            cursor: pointer;
-            font-weight: bold;
-        }
-
         #ingre-list button{
             background-color: transparent;
             border: rgb(34,139,34) 1px solid;
@@ -232,7 +202,7 @@ class ViewMyRecipe extends HTMLElement {
             display: flex;
             flex-flow: row nowrap;
             justify-content: flex-end;
-            // margin: 90px 60px 60px 60px;
+            margin: 90px 60px 60px 60px;
             width: 100%;
         }
 
@@ -254,6 +224,7 @@ class ViewMyRecipe extends HTMLElement {
             color: white;
         }
 
+
         /* aside */
         article > aside {
             align-items: center;
@@ -261,12 +232,11 @@ class ViewMyRecipe extends HTMLElement {
             display: flex;
             flex: 1 1 25%;
             flex-flow: column nowrap;
-            // padding-bottom: 60px;
+            padding-bottom: 60px;
         }
 
-    
         #ingre-aside {
-            margin: 15px 15px;
+            margin: 40px 15px;
         }
 
         #ingre-aside h2 {
@@ -294,24 +264,6 @@ class ViewMyRecipe extends HTMLElement {
             width: 20px;
             height: 20px;
         }
-
-        #show-ingre-btn {
-            display: none;
-        }
-        @media (max-width: 1091px) {
-            #show-ingre-btn {
-                width: 100%;
-                height: 70px;
-                display: block;
-                background-color: white;
-                object-fit: contain;
-            }
-
-            #show-ingre-btn:hover {
-                cursor: pointer;
-            }
-        }
-        
         `;
         styleElem.innerHTML = styles;
 
@@ -374,7 +326,7 @@ class ViewMyRecipe extends HTMLElement {
 
         let cookTime = document.createElement("p");
         cookTime.textContent = `${getTime(data)} min`;
-        
+
         let speechSynthesis = window.speechSynthesis;
         let textToSpeech = document.createElement("button");
         textToSpeech.setAttribute("id", "tts-btn");
@@ -421,8 +373,8 @@ class ViewMyRecipe extends HTMLElement {
         mainFooterSec.appendChild(backButton);
         mainFooterSec.appendChild(editButton);
 
-        // main.appendChild(mainHeaderSec);
-        // main.appendChild(leftMainSec);
+        main.appendChild(mainHeaderSec);
+        main.appendChild(leftMainSec);
         main.appendChild(rightMainSec);
         main.appendChild(mainFooterSec);
 
@@ -533,33 +485,12 @@ class ViewMyRecipe extends HTMLElement {
         dietAside.appendChild(glutenSec);
         dietAside.appendChild(vegetaSec);
 
-        // aside.appendChild(mainHeaderSec);
-        // aside.appendChild(leftMainSec);
         aside.appendChild(ingreAside);
         aside.appendChild(dietAside);
 
-        let showIngre = document.createElement("img");
-        showIngre.setAttribute("id", "show-ingre-btn");
-        showIngre.setAttribute("src", "./assets/images/icons/arrow-up.png");
-        showIngre.addEventListener("click", () => {
-            if (ingreAside.style.display != "none" && dietAside.style.display != "none") {  
-                ingreAside.style.display = "none";
-                dietAside.style.display = "none";
-                showIngre.setAttribute("src", "./assets/images/icons/arrow-down.png");
-            } else {
-                ingreAside.style.display = "block";
-                dietAside.style.display = "block";
-                showIngre.setAttribute("src", "./assets/images/icons/arrow-up.png");
-            }
-        });
-
-        aside.appendChild(showIngre)
-
         card.appendChild(header);
-        card.appendChild(mainHeaderSec);
-        card.appendChild(leftMainSec);  
-        card.appendChild(aside);
         card.appendChild(main);
+        card.appendChild(aside);
 
         this.shadow.appendChild(styleElem);
         this.shadow.appendChild(card);
@@ -599,19 +530,20 @@ class ViewMyRecipe extends HTMLElement {
         $("#add-recipe-page").classList.add("main-shown");
     }
 
+
     /**
      * Play the instruction step by step
      * @param{Object} speechSynthesis a speech Object 
      * @returns Void
      */
-     playTextToSpeech(speechSynthesis) {
+    playTextToSpeech(speechSynthesis) {
         let recipeText = this.shadowRoot.querySelector("#steps-list");
         recipeText = recipeText.querySelectorAll("li");
         let i = 0;
         let speechText = new SpeechSynthesisUtterance(recipeText[i].textContent);
         speechSynthesis.speak(speechText);
         let myRecipeView = document.querySelector("#view-recipe-page").children[0];
-        myRecipeView.addEventListener('keydown', function(event) {
+        myRecipeView.addEventListener('keydown', function (event) {
             console.log("MyRecipePage");
             if (event.key == "ArrowRight" && i < recipeText.length - 1) {
                 i++;
@@ -631,7 +563,7 @@ class ViewMyRecipe extends HTMLElement {
                 speechSynthesis.cancel();
                 speechSynthesis.speak(speechText);
             }
-        });         
+        });
     }
 }
