+++ source/assets/components/ViewMyRecipeMobile.js
@@ -1,13 +1,12 @@
 import { $, loadMain, router } from "../scripts/main.js";
-import { checkFav, rmFav, addFav } from "../scripts/helpCrudFunc.js";
-import { getImgUrl, getTitle, getTime, getFeaturedSteps, getIngreFea } from "../scripts/helpGetDataFunc.js";
+import { getImgUrl, getTitle, getTime, getSteps, getIngre } from "../scripts/helpGetDataFunc.js";
 import { getDairy, getGluten, getVegan, getVegeta } from "../scripts/helpGetDataFunc.js";
 
 /**
- * This is the component for the View Featured Recipe Page.
+ * This is the component for the View My Recipe Page.
  * @class
  */
-class ViewFeaRecipe extends HTMLElement {
+class ViewMyRecipeMobile extends HTMLElement {
     /**
      * Attach the shadowroot which contains the View Recipe Page materials.
      * @constructor
@@ -38,8 +37,7 @@ class ViewFeaRecipe extends HTMLElement {
             flex-flow: row wrap;
             margin: 10px 0px 90px 0px;
         }
-
-
+        
         /* header */
         article > header {
             align-items: center;
@@ -49,6 +47,7 @@ class ViewFeaRecipe extends HTMLElement {
             height: 12vh;
             justify-content: space-between;
             width: 100%;
+            border-style:dotted;
         }
 
         #logo-sec {
@@ -59,7 +58,7 @@ class ViewFeaRecipe extends HTMLElement {
 
         #logo-sec > a > img {
             height: 100%;
-            width: 125px;
+            width: 100px;
             object-fit: cover;
         }
 
@@ -83,9 +82,25 @@ class ViewFeaRecipe extends HTMLElement {
         #home-link {
             color: white;
             font-size: 2vw;
-            margin: 0px 0px 0px 20px;
+            // margin: 0px 0px 0px 20px;
             text-decoration: none;
         }
+
+        @media (max-width: 480px) {
+            article {
+                box-shadow: 0px 0px 15px #888888;
+                display: flex;
+                flex-flow: row wrap;
+                margin: 10px 0px 90px 0px;
+            }
+
+            #header-title {
+                font-size: 5vw; 
+                position:relative;
+                left: 6vw;
+            }
+
+          }
 
 
         /* main */
@@ -93,7 +108,7 @@ class ViewFeaRecipe extends HTMLElement {
             align-content: flex-start;
             display: flex;
             flex: 1 1 70%;
-            flex-flow: row wrap;
+            flex-flow: column wrap;
             justify-content: center;
             padding: 15px;
         }
@@ -102,30 +117,34 @@ class ViewFeaRecipe extends HTMLElement {
         #main-header {
             align-items: center;
             display: flex;
-            flex-flow: row nowrap;
+            flex-flow: column nowrap;
             justify-content: center;
-            margin: 20px 0px 50px 0px;
+            margin: 20px 8px;
             width: 100%;
+            text-align: center;
         }
 
         #main-header > h1 {
+            width: 80%;
             line-height: 35px;
-            font-size: 30px;
+            font-size: 23px;
             margin: 0px 30px;
-            width: 50%;
         }
 
-        #main-header > img {
-            width: 25px;
-            height: 25px;
+        @media (max-width: 480px) {
+            article > main {
+                width: calc(100vw - 480px);
+            }
         }
+
 
         /* left-main */
         #left-main {
             align-items: center;
             display: flex;
-            flex: 1 1 40%;
+            // flex: 1 1 40%;
             flex-flow: column nowrap;
+            width: 100%;
         }
 
         #left-main > img {
@@ -134,12 +153,14 @@ class ViewFeaRecipe extends HTMLElement {
         }
 
         #left-main > h2 {
-            margin: 60px 0px 20px 0px;
+            display: none;
+            // margin: 10px 0px 20px 0px;
         }
         
         #left-main > p {
+            visibility: hidden;
             font-size: 20px;
-            margin-bottom: 40px;
+            // margin-bottom: 10px;
         }
 
         #left-main > button {
@@ -153,10 +174,6 @@ class ViewFeaRecipe extends HTMLElement {
             padding: 5px 20px;
         }
         
-        #tts-btn {
-            margin-top: 10px;
-        }
-
         #left-main > button:hover {
             border: 1px solid #313131;
             background: darkgreen;
@@ -192,13 +209,30 @@ class ViewFeaRecipe extends HTMLElement {
             font-weight: bold;
         }
 
+        #ingre-list button{
+            background-color: transparent;
+            border: rgb(34,139,34) 1px solid;
+            border-radius: 4px;
+            color: rgb(34,139,34);
+            
+            cursor: pointer;
+            font-size: 12px;
+            
+            padding: .2em .8em;
+            margin: 0 .2em;
+        }
+        #ingre-list button:hover{
+            background: white;
+            color: #305a50;
+        }
+
         /* main-footer */
         #main-footer{
             align-items: center;
             display: flex;
             flex-flow: row nowrap;
-            justify-content: flex-end;
-            margin: 90px 60px 60px 60px;
+            justify-content: center;
+            // margin: 90px 60px 60px 60px;
             width: 100%;
         }
 
@@ -209,7 +243,7 @@ class ViewFeaRecipe extends HTMLElement {
             color: #305a50;
             cursor: pointer;
             font-size: 16px;
-            margin: 0px 20px 0px 0px;
+            margin: 20px 5px;
             min-width: 120px;
             padding: 5px 20px;
         }
@@ -220,7 +254,6 @@ class ViewFeaRecipe extends HTMLElement {
             color: white;
         }
 
-
         /* aside */
         article > aside {
             align-items: center;
@@ -228,11 +261,12 @@ class ViewFeaRecipe extends HTMLElement {
             display: flex;
             flex: 1 1 25%;
             flex-flow: column nowrap;
-            padding-bottom: 60px;
+            margin-top: 20px;
         }
 
+    
         #ingre-aside {
-            margin: 40px 15px;
+            margin: 15px 15px;
         }
 
         #ingre-aside h2 {
@@ -260,6 +294,24 @@ class ViewFeaRecipe extends HTMLElement {
             width: 20px;
             height: 20px;
         }
+
+        #show-ingre-btn {
+            display: none;
+        }
+        @media (max-width: 1091px) {
+            #show-ingre-btn {
+                width: 100%;
+                height: 70px;
+                display: block;
+                background-color: white;
+                object-fit: contain;
+            }
+
+            #show-ingre-btn:hover {
+                cursor: pointer;
+            }
+        }
+        
         `;
         styleElem.innerHTML = styles;
 
@@ -272,7 +324,6 @@ class ViewFeaRecipe extends HTMLElement {
         logoSec.setAttribute("id", "logo-sec");
         titleSec.setAttribute("id", "title-sec");
         holderSec.setAttribute("id", "holder-sec");
-
 
         let header = document.createElement("header");
         let headerHomeLink = document.createElement("a");
@@ -307,20 +358,8 @@ class ViewFeaRecipe extends HTMLElement {
         let recipeTitle = document.createElement("h1");
         recipeTitle.textContent = getTitle(data);
 
-        let heartImg = document.createElement("img");
-        if (checkFav(getTitle(data))) {
-            heartImg.setAttribute("src", "assets/images/icons/fillHeart.svg");
-        }
-        else {
-            heartImg.setAttribute("src", "assets/images/icons/emptyHeart.svg");
-        }
-
-        heartImg.addEventListener("click", () => {
-            this.changeHeart(data, heartImg);
-        });
 
         mainHeaderSec.appendChild(recipeTitle);
-        mainHeaderSec.appendChild(heartImg);
 
         // left-main
         let leftMainSec = document.createElement("section");
@@ -336,12 +375,6 @@ class ViewFeaRecipe extends HTMLElement {
         let cookTime = document.createElement("p");
         cookTime.textContent = `${getTime(data)} min`;
 
-        let toNutritionPage = document.createElement("button");
-        toNutritionPage.textContent = "Nutrition Facts";
-        toNutritionPage.addEventListener("click", () => {
-            this.viewNutrition(data);
-        })
-
         let speechSynthesis = window.speechSynthesis;
         let textToSpeech = document.createElement("button");
         textToSpeech.setAttribute("id", "tts-btn");
@@ -353,7 +386,7 @@ class ViewFeaRecipe extends HTMLElement {
         leftMainSec.appendChild(recipeImg);
         leftMainSec.appendChild(timeLabel);
         leftMainSec.appendChild(cookTime);
-        leftMainSec.appendChild(toNutritionPage);
+        // leftMainSec.appendChild(toNutPage);
         leftMainSec.appendChild(textToSpeech);
 
         // right-main
@@ -362,34 +395,10 @@ class ViewFeaRecipe extends HTMLElement {
 
         let stepsTitle = document.createElement("h2");
         stepsTitle.textContent = "Procedure and steps: ";
-        // stepsSec.innerHTML = getSteps(data);
 
         let stepsSec = document.createElement("section");
-        // stepsSec.innerHTML = getSteps(data);
         stepsSec.id = "steps-list";
-        let instructionArray = getFeaturedSteps(data);
-        let instructionOrderedList = document.createElement("ol");
-
-        // Loop through steps within the analyzedInstructions index
-        for (let i = 0; i < instructionArray.length; i++) {
-            let instructionItem = instructionArray[i]["steps"];
-            for (let j = 0; j < instructionItem.length; j++) {
-                let instructionStep = document.createElement("li");
-                instructionStep.innerHTML = `Step ${j + 1}: ` + instructionItem[j]["step"];
-                instructionStep.addEventListener("click", () => {
-                    let speechText = new SpeechSynthesisUtterance(instructionStep.innerHTML);
-                    speechSynthesis.cancel();
-                    speechSynthesis.speak(speechText);
-                })
-                instructionOrderedList.appendChild(instructionStep);
-            }
-        }
-
-        stepsSec.appendChild(instructionOrderedList);
-
-        if (stepsSec.innerHTML == '') {
-            stepsSec.innerHTML = 'OOOPS! The recipe does not contain any procedure or steps. Please start using your imagination!'
-        }
+        stepsSec.innerHTML = getSteps(data);
 
         rightMainSec.appendChild(stepsTitle);
         rightMainSec.appendChild(stepsSec);
@@ -400,12 +409,20 @@ class ViewFeaRecipe extends HTMLElement {
 
         let backButton = document.createElement("button");
         backButton.textContent = "Back";
-        backButton.addEventListener("click", this.feaRecipeToLand);
+        backButton.addEventListener("click", this.myRecipeToLand);
+
+        let editButton = document.createElement("button");
+        editButton.textContent = "Edit";
+        editButton.addEventListener("click", () => {
+            this.myRecipeToEdit(data);
+        });
+
 
         mainFooterSec.appendChild(backButton);
+        mainFooterSec.appendChild(editButton);
 
-        main.appendChild(mainHeaderSec);
-        main.appendChild(leftMainSec);
+        // main.appendChild(mainHeaderSec);
+        // main.appendChild(leftMainSec);
         main.appendChild(rightMainSec);
         main.appendChild(mainFooterSec);
 
@@ -422,8 +439,7 @@ class ViewFeaRecipe extends HTMLElement {
 
         let ingreListSec = document.createElement("section");
         ingreListSec.id = "ingre-list";
-        ingreListSec.innerHTML = getIngreFea(data);
-
+        ingreListSec.innerHTML = getIngre(data);
 
         ingreAside.appendChild(ingreLabel);
         ingreAside.appendChild(ingreListSec);
@@ -517,92 +533,51 @@ class ViewFeaRecipe extends HTMLElement {
         dietAside.appendChild(glutenSec);
         dietAside.appendChild(vegetaSec);
 
+        // aside.appendChild(mainHeaderSec);
+        // aside.appendChild(leftMainSec);
         aside.appendChild(ingreAside);
         aside.appendChild(dietAside);
 
+        let showIngre = document.createElement("img");
+        showIngre.setAttribute("id", "show-ingre-btn");
+        showIngre.setAttribute("src", "./assets/images/icons/arrow-up.png");
+        showIngre.addEventListener("click", () => {
+            if (ingreAside.style.display != "none" && dietAside.style.display != "none") {
+                ingreAside.style.display = "none";
+                dietAside.style.display = "none";
+                showIngre.setAttribute("src", "./assets/images/icons/arrow-down.png");
+            } else {
+                ingreAside.style.display = "block";
+                dietAside.style.display = "block";
+                showIngre.setAttribute("src", "./assets/images/icons/arrow-up.png");
+            }
+        });
+
+        aside.appendChild(showIngre)
+
         card.appendChild(header);
-        card.appendChild(main);
+        card.appendChild(mainHeaderSec);
+        card.appendChild(leftMainSec);
         card.appendChild(aside);
+        card.appendChild(main);
 
         this.shadow.appendChild(styleElem);
         this.shadow.appendChild(card);
-        // console.log(ingreListSec.querySelector("li[id='1']").querySelector("p[id='amount']").getAttribute("unit"));
-        let ingredientsOL = ingreListSec.querySelector("ol").querySelectorAll("li");
-
-        // Loop thru ol and parse the necessary parameters
-        for (let i = 0; i < ingredientsOL.length; i++) {
-            let requestBody = {};
-            let dropdown = ingredientsOL[i].querySelector("select");
-            requestBody["ingredientName"] = ingredientsOL[i].querySelector("p[id='name']").getAttribute("name");
-            requestBody["sourceAmount"] = ingredientsOL[i].querySelector("p[id='amount']").getAttribute("amount");
-            requestBody["sourceUnit"] = ingredientsOL[i].querySelector("p[id='amount']").getAttribute("unit");
-
-            dropdown.addEventListener("change", async () => {
-                requestBody["targetUnit"] = dropdown.options[dropdown.selectedIndex].value;
-                if (dropdown.options[dropdown.selectedIndex].value == "Select") {
-                    requestBody["targetUnit"] = dropdown.options[dropdown.selectedIndex].value = "";
-                }
-                let convertInit = async (dataToConvert, locationObject) => {
-                    let convertSuccess = await this.fetchConvertUnit(dataToConvert, locationObject);
-                    if (!convertSuccess) {
-                        console.log("Convert Sucess");
-                        return;
-                    }
-                }
-                convertInit(requestBody, ingredientsOL[i]); // Call the netlify function for API call
-            });
-        }
     }
 
     /**
-     * Toggles on and off the heart based on favorite.
-     * @param {Object} data a JSON data object contains information to keep track of which recipe is being added to favorites.
-     * @param {HTMLElement} cardObj an HTML element that contains the heart image.
-     * @return Void
-     */
-    changeHeart(data, cardObj) {
-        if (checkFav(getTitle(data))) {
-            cardObj.setAttribute("src", "assets/images/icons/emptyHeart.svg");
-            rmFav(getTitle(data));
-        }
-        else {
-            cardObj.setAttribute("src", "assets/images/icons/fillHeart.svg");
-            addFav(data);
-        }
-    }
-
-    /**
-     * View nutrition facts about the selected recipe.
-     * @param {Object} data a JSON data object contains information to load the nutrition page.
+     * Leave View My Recipe Page to landing page.
      * @returns Void
      */
-    viewNutrition(data) {
-        $("#view-recipe-page").classList.remove("main-shown");
-        $("#view-recipe-page").innerHTML = "";
-        speechSynthesis.cancel();
-        $("#view-nutrition-page").classList.add("main-shown");
-        const nutritionPage = document.createElement("nutrition-page");
-        nutritionPage.data = data;
-        $("#view-nutrition-page").appendChild(nutritionPage);
-    }
-
-
-    /**
-     * Leave Featured Recipe Page to landing page using router object.
-     * @returns Void
-     */
-    feaRecipeToLand() {
+    myRecipeToLand() {
         $("#view-recipe-page").classList.remove("main-shown");
         $("#view-recipe-page").innerHTML = "";
         speechSynthesis.cancel();
         loadMain();
-        if ($("#featured-page").classList.contains("shown")) {
-            router.navigate("ToFeaturedPage");
+        if ($("#my-page").classList.contains("shown")) {
+            router.navigate("ToMyRecipePage");
         }
-        else if ($("#favorite-page").classList.contains("shown")) {
-            router.navigate("ToFavoritePage");
-        }
-        else if ($("#search-featured").classList.contains("shown")) {
+        else if ($("#search-my").classList.contains("shown")) {
             router.navigate("ToSearchPage");
         }
         else {
@@ -610,6 +585,19 @@ class ViewFeaRecipe extends HTMLElement {
         }
     }
 
+    /**
+     * Leave View My Recipe Page to edit page.
+     * @param {Object} data a JSON data object contains information to populate this component.
+     * @returns Void
+     */
+    myRecipeToEdit(data) {
+        $("#view-recipe-page").classList.remove("main-shown");
+        $("#view-recipe-page").innerHTML = "";
+        let editRecipePage = document.createElement("edit-recipe-mobile");
+        editRecipePage.data = data;
+        $("#add-recipe-page").appendChild(editRecipePage);
+        $("#add-recipe-page").classList.add("main-shown");
+    }
 
     /**
      * Play the instruction step by step
@@ -622,9 +610,9 @@ class ViewFeaRecipe extends HTMLElement {
         let i = 0;
         let speechText = new SpeechSynthesisUtterance(recipeText[i].textContent);
         speechSynthesis.speak(speechText);
-        let featuredView = document.querySelector("#view-recipe-page").children[0];
-        featuredView.addEventListener('keydown', function (event) {
-            console.log("FeaRecipePage");
+        let myRecipeView = document.querySelector("#view-recipe-page").children[0];
+        myRecipeView.addEventListener('keydown', function (event) {
+            console.log("MyRecipePage");
             if (event.key == "ArrowRight" && i < recipeText.length - 1) {
                 i++;
                 let speechText = new SpeechSynthesisUtterance(recipeText[i].textContent);
@@ -645,31 +633,8 @@ class ViewFeaRecipe extends HTMLElement {
             }
         });
     }
-
-    /**
-     * Make a GET call to the netlify function using provided parameters, which would then
-     * returns the converted result from spoonacular API.
-     * @param {Object} dataToConvert JSON object contains parsed unit conversion data.
-     * @param {HTMLElement} locationObject HTML element contain the location to display result.
-     * @returns a Promise of fetched data.
-     */
-    async fetchConvertUnit(dataToConvert, locationObject) {
-        return new Promise((resolve, reject) => {
-            fetch("./.netlify/functions/convert-unit?" + new URLSearchParams({
-                ingredientName: dataToConvert.ingredientName,
-                sourceAmount: dataToConvert.sourceAmount,
-                sourceUnit: dataToConvert.sourceUnit,
-                targetUnit: dataToConvert.targetUnit
-            }))
-                .then((response) => response.json())
-                .then((data) => {
-                    locationObject.querySelector("span[id='converted-result']").innerHTML = data.targetAmount + data.targetUnit;
-                    console.log(locationObject.querySelector("span[id='converted-result']").innerHTML);
-                    resolve(true);
-                }).catch(() => reject(false));
-        });
-    }
 }
 
+
 // Define the "view-fea-recipe" element using this class.
-customElements.define("view-fea-recipe", ViewFeaRecipe);
+customElements.define("view-my-recipe-mobile", ViewMyRecipeMobile);
