+++ source/index.html
@@ -16,11 +16,15 @@
     <script src="assets/components/EmptyCard.js" type="module"></script>
     <script src="assets/components/RecipeCardFeaturedPG.js" type="module"></script>
     <script src="assets/components/ViewFeaRecipe.js" type="module"></script>
+    <script src="assets/components/ViewFeaRecipeMobile.js" type="module"></script>
     <script src="assets/components/ViewMyRecipe.js" type="module"></script>
+    <script src="assets/components/ViewMyRecipeMobile.js" type="module"></script>
     <script src="assets/components/AddRecipe.js" type="module"></script>
+    <script src="assets/components/AddRecipeMobile.js" type="module"></script>
     <script src="assets/components/RecipeCardMyMyPage.js" type="module"></script>
     <script src="assets/components/NewCardMyPage.js" type="module"></script>
     <script src="assets/components/EditRecipe.js" type="module"></script>
+    <script src="assets/components/EditRecipeMobile.js" type="module"></script>
     <script src="assets/components/deleteConfirmation.js" type="module"></script>
     <script src="assets/components/NutritionPage.js" type="module"></script>
 
@@ -67,24 +71,24 @@
                         <input type="checkbox" id="vegan-filter">
                         <label for="vegan-filter">Vegan</label>
                     </div>
-                   
+
                     <div id="dairy-filter-div">
                         <input type="checkbox" id="dairy-filter">
                         <label for="dairy-filter">Dairy Free</label>
                     </div>
-    
+
                     <div id="gluten-filter-div">
                         <input type="checkbox" id="gluten-filter">
                         <label for="gluten-filter">Gluten Free</label>
                     </div>
-                    
+
                     <div id="vegetarian-filter-div">
                         <input type="checkbox" id="vegetarian-filter">
                         <label for="vegetarian-filter">Vegetarian</label>
                     </div>
                 </div>
             </details>
-            
+
         </section>
 
         <section class="main-item shown" id="featured-recipes">
